refactor: split core/mod.rs into focused sub-modules, drop re-exports
Break the 1,455-line StitchCore orchestrator into 3 new files:

- core/render.rs (390 lines): all submit_frame_* and render_*_at_pose
- core/replay_management.rs (290 lines): stacked replay wiring + GPU pack
- core/pose.rs (140 lines): pose resolution, detection helpers

core/mod.rs retains struct definition, constructor, wiring, accessors,
and tests (652 lines, 503 non-test).

Removed pub use re-exports from core/mod.rs. Consumers now import
from real module paths (e.g., reco_core::core::types::StitchCoreConfig).
Updated reco-obs, reco-io, session, reco-control import paths.

8 tests pass, clippy clean.
