+++ crates/reco-core/src/core/mod.rs
@@ -33,53 +33,33 @@
 //!
 //! - [`types`] - error types, config, render outcome, replay frame, recorder traits
 //! - [`replay_buffer`] - bounded-duration ring buffer for replay frames
-//!
-//! ## Usage (push, live)
-//!
-//! ```rust,ignore
-//! use reco_core::core::{StitchCore, StitchCoreConfig, RenderOutcome};
-//!
-//! let mut core = StitchCore::new(gpu, StitchCoreConfig {
-//!     calibration,
-//!     viewport,
-//!     input_width, input_height,
-//!     output_format: wgpu::TextureFormat::Rgba8Unorm,
-//!     input_format: InputFormat::Yuv420p,
-//!     ..Default::default()
-//! });
-//!
-//! core.set_director(Box::new(MyDirector::new()));
-//!
-//! // Per-frame, from the consumer's callback thread:
-//! match core.submit_frame_yuv(&left, &right)? {
-//!     RenderOutcome::Rgba(bytes) => compositor.upload(bytes),
-//!     RenderOutcome::Warmup => { /* first two frames: pipeline filling */ }
-//! }
-//! ```
+//! - [`render`] - submit and render-at-pose methods
+//! - [`replay_management`] - stacked replay recorder wiring (CPU + GPU paths)
+//! - [`pose`] - pose resolution, detection scheduling, panorama mapping
 
+mod pose;
+mod render;
 pub mod replay_buffer;
+mod replay_management;
 pub mod types;
-
-pub use replay_buffer::ReplayBuffer;
-pub use types::*;
 
 use std::time::{Duration, Instant};
 
 use crate::calibration::MatchCalibration;
-use crate::detector::{
-    CameraId, ChromaFormat, Detection, DetectorFrame, RawFrame, UnifiedDetector,
-};
+use crate::detector::UnifiedDetector;
 use crate::director::{MappedDetection, ViewportPosition};
 use crate::gpu::GpuContext;
 use crate::panner::Panner;
-use crate::pipeline::{BgraPlanes, StitchPipeline, YuvPlanes};
-use crate::projection::{self, CoverageBoundary, LShapeProjection, PanoramaExtent, Projection};
-use crate::renderer::InputFormat;
+use crate::pipeline::StitchPipeline;
+use crate::projection::{CoverageBoundary, LShapeProjection, PanoramaExtent, Projection};
 use crate::rgba_readback::RgbaReadback;
 use crate::source::{CameraInput, StereoCameraInput};
 use crate::tracker::Tracker;
-use crate::yuv_stack_packer::{
-    OutputTileSize, SourceFormat, StackGridLayout, StackedPackSource, YuvStackPacker,
+use crate::yuv_stack_packer::YuvStackPacker;
+
+use self::replay_buffer::ReplayBuffer;
+use self::types::{
+    StackedReplayGpuRecorder, StackedReplayRecorder, StitchCoreConfig, StitchCoreError,
 };
 
 /// Canonical push-first stitching core.
@@ -103,14 +83,14 @@ use crate::yuv_stack_packer::{
 /// attached director each submit; directors see a non-empty
 /// `detections` slice on detection frames, empty otherwise.
 pub struct StitchCore {
-    pipeline: StitchPipeline,
-    readback: RgbaReadback,
-    output_width: u32,
-    output_height: u32,
+    pub(crate) pipeline: StitchPipeline,
+    pub(crate) readback: RgbaReadback,
+    pub(crate) output_width: u32,
+    pub(crate) output_height: u32,
 
-    projection: Box<dyn Projection>,
-    camera_input: Box<dyn CameraInput>,
-    coverage: Option<CoverageBoundary>,
+    pub(crate) projection: Box<dyn Projection>,
+    pub(crate) camera_input: Box<dyn CameraInput>,
+    pub(crate) coverage: Option<CoverageBoundary>,
 
     /// Per-class trackers that feed a shared [`WorldState`](crate::tracker::WorldState)
     /// consumed by [`StitchCore::panner`]. Slot-based on purpose:
@@ -120,28 +100,28 @@ pub struct StitchCore {
     /// The panner only runs when at least one tracker is registered
     /// AND a panner is set. Otherwise the pose stays at the pipeline
     /// default.
-    ball_tracker: Option<Box<dyn Tracker>>,
-    player_tracker: Option<Box<dyn Tracker>>,
+    pub(crate) ball_tracker: Option<Box<dyn Tracker>>,
+    pub(crate) player_tracker: Option<Box<dyn Tracker>>,
     /// Camera-motion policy. Consumes the assembled
     /// [`WorldState`](crate::tracker::WorldState) each frame and emits
     /// a [`ViewportPosition`]. When unset, the pose stays at the
     /// pipeline default.
-    panner: Option<Box<dyn Panner>>,
+    pub(crate) panner: Option<Box<dyn Panner>>,
     /// Previous frame's resolved pose, passed to the panner in its
     /// [`PanContext`](crate::panner::PanContext) so panners can
     /// compute first-order motion deltas statelessly.
-    previous_panner_pose: ViewportPosition,
+    pub(crate) previous_panner_pose: ViewportPosition,
 
-    detector: Option<Box<dyn UnifiedDetector>>,
+    pub(crate) detector: Option<Box<dyn UnifiedDetector>>,
     /// How often detection runs. 1 = every frame (default), higher =
     /// skip frames. On skipped frames the director still ticks with
     /// the previously tracked detections.
-    detection_interval: u64,
+    pub(crate) detection_interval: u64,
     /// Panorama-mapped detections from the last detection frame.
     /// Reused on skipped frames so the director retains context.
-    last_detections: Vec<MappedDetection>,
+    pub(crate) last_detections: Vec<MappedDetection>,
 
-    replay: Option<ReplayBuffer>,
+    pub(crate) replay: Option<ReplayBuffer>,
 
     /// Optional stacked-video replay recorder attached via
     /// [`Self::set_stacked_recorder`]. Fires on every successful
@@ -149,13 +129,13 @@ pub struct StitchCore {
     /// Decouples reco-core from the actual encoder implementation
     /// (lives in reco-io under `stacked-output`) so mobile / wasm
     /// builds that skip reco-io see no replay-recording code.
-    stacked_recorder: Option<Box<dyn StackedReplayRecorder>>,
+    pub(crate) stacked_recorder: Option<Box<dyn StackedReplayRecorder>>,
 
     /// Optional GPU-pack packer attached via
     /// [`Self::enable_gpu_stacked_replay`]. Holds the compute
     /// pipelines and triple-buffered staging ring. `None` when the
     /// session runs on a CPU-pack (or no replay) path.
-    stacked_packer: Option<YuvStackPacker>,
+    pub(crate) stacked_packer: Option<YuvStackPacker>,
 
     /// Optional GPU-pack atlas recorder attached via
     /// [`Self::set_stacked_gpu_recorder`]. Receives the packed atlas
@@ -163,7 +143,7 @@ pub struct StitchCore {
     /// completed readback slot. `None` means the pack still runs
     /// (if enabled) but the bytes are dropped - useful when a
     /// consumer wants to attach the recorder lazily.
-    stacked_gpu_recorder: Option<Box<dyn StackedReplayGpuRecorder>>,
+    pub(crate) stacked_gpu_recorder: Option<Box<dyn StackedReplayGpuRecorder>>,
 
     /// Whether `resolve_current_pose` clamps output through the
     /// coverage boundary (FRICTION A13 - "constrained look"). `true`
@@ -175,10 +155,10 @@ pub struct StitchCore {
     /// The public [`Self::safe_clamp`] method remains available
     /// regardless of this flag - it's the primitive consumers use
     /// for ad-hoc clamping outside the render loop.
-    constrained_look: bool,
+    pub(crate) constrained_look: bool,
 
-    frame_count: u64,
-    session_start: Option<Instant>,
+    pub(crate) frame_count: u64,
+    pub(crate) session_start: Option<Instant>,
 }
 
 impl StitchCore {
@@ -242,384 +222,6 @@ impl StitchCore {
             frame_count: 0,
             session_start: None,
         })
-    }
-
-    // -----------------------------------------------------------------
-    // Submit / render
-    // -----------------------------------------------------------------
-
-    /// Submit a stereo YUV420P frame pair and render the current pose.
-    ///
-    /// Uses the director (if attached) and coverage clamping to pick
-    /// the viewport, renders, and reads back RGBA. The first two calls
-    /// produce [`RenderOutcome::Warmup`] while the triple-buffered
-    /// staging ring fills; from the third call onward every submit
-    /// yields RGBA bytes from two frames ago.
-    pub fn submit_frame_yuv(
-        &mut self,
-        left: &YuvPlanes<'_>,
-        right: &YuvPlanes<'_>,
-    ) -> Result<RenderOutcome<'_>, StitchCoreError> {
-        self.anchor_session_start();
-
-        // Feed the stacked-video replay recorder before render so
-        // the recording captures the exact planes the pipeline will
-        // see. Errors inside the recorder are logged by the impl;
-        // never propagate them - a failing recorder must not break
-        // the live stitch output.
-        if let Some(ref mut recorder) = self.stacked_recorder {
-            let (src_w, src_h) = self.pipeline.source_info();
-            recorder.record_yuv(left, right, src_w, src_h);
-        }
-
-        // Detection first, so the director's `update` tick in
-        // resolve_current_pose sees the latest tracked objects. Skipped
-        // frames reuse last_detections so the director still has context.
-        let ran_detection = self.detector.is_some() && self.should_run_detection();
-        if ran_detection {
-            let (src_w, src_h) = self.pipeline.source_info();
-            let dets = self.run_yuv_detection(left, right, src_w, src_h);
-            self.last_detections = self.map_detections_to_panorama(dets);
-        }
-
-        let pose = self.resolve_current_pose(ran_detection);
-        let cmd = self
-            .pipeline
-            .render_to_target(left, right, pose.yaw, pose.pitch)?;
-        // GPU stacked-replay pack runs before the readback so the
-        // borrow checker sees `self.readback` free while the pack
-        // runs. Queue ordering: `queue.write_texture` inside
-        // `render_to_target` is already enqueued; the pack submit
-        // processes the writes before its compute pass reads the
-        // textures, and the subsequent stitch submit reads the
-        // same textures into the render target. No-op when packer
-        // is not enabled.
-        self.drive_gpu_stacked_pack();
-        // Split-borrow: push_replay only accesses self.replay +
-        // self.session_start; self.readback keeps the rgba slice
-        // alive. Inlining the replay push (instead of going through
-        // `&mut self` on a helper) lets the borrow checker see the
-        // fields are disjoint.
-        let captured_at = self.session_start.map(|s| s.elapsed()).unwrap_or_default();
-        let rgba =
-            self.readback
-                .readback(self.pipeline.gpu(), self.pipeline.render_target(), cmd)?;
-        self.frame_count += 1;
-        if let (Some(replay), Some(bytes)) = (self.replay.as_mut(), rgba) {
-            replay.push(ReplayFrame {
-                rgba: bytes.to_vec(),
-                captured_at,
-                pose,
-            });
-        }
-        Ok(match rgba {
-            Some(bytes) => RenderOutcome::Rgba(bytes),
-            None => RenderOutcome::Warmup,
-        })
-    }
-
-    /// Submit a stereo YUV420P frame pair at an explicit pose.
-    ///
-    /// Same full loop as [`Self::submit_frame_yuv`] - anchors the
-    /// session-start clock, runs detection when
-    /// `frame_count % detection_interval == 0`, renders, reads back
-    /// RGBA, pushes into the replay buffer, increments frame_count -
-    /// but bypasses the director and uses the caller-supplied
-    /// `(yaw, pitch)` directly. The FOV stays at whatever the
-    /// pipeline currently has (set via [`Self::pipeline_mut`] or
-    /// `update_calibration`).
-    ///
-    /// This is the canonical submit path for interactive UIs (OBS
-    /// pan/zoom sliders, mouse-drag preview) where pose comes from
-    /// user input rather than a director.
-    pub fn submit_frame_yuv_at_pose(
-        &mut self,
-        left: &YuvPlanes<'_>,
-        right: &YuvPlanes<'_>,
-        yaw: f32,
-        pitch: f32,
-    ) -> Result<RenderOutcome<'_>, StitchCoreError> {
-        self.anchor_session_start();
-
-        // Replay recording tap - see `submit_frame_yuv` for the
-        // rationale (record-before-render so the file exactly
-        // matches what the pipeline consumed).
-        if let Some(ref mut recorder) = self.stacked_recorder {
-            let (src_w, src_h) = self.pipeline.source_info();
-            recorder.record_yuv(left, right, src_w, src_h);
-        }
-
-        // `submit_frame_yuv_at_pose` bypasses resolve_current_pose (caller
-        // provides the pose directly), but detection still runs on the
-        // schedule so directors stay populated for a later `current_pose()`
-        // peek or a regular `submit_frame_yuv` submit.
-        if self.detector.is_some() && self.should_run_detection() {
-            let (src_w, src_h) = self.pipeline.source_info();
-            let dets = self.run_yuv_detection(left, right, src_w, src_h);
-            self.last_detections = self.map_detections_to_panorama(dets);
-        }
-
-        let cmd = self.pipeline.render_to_target(left, right, yaw, pitch)?;
-        // GPU stacked-replay pack - see `submit_frame_yuv` for
-        // ordering rationale. No-op when not enabled.
-        self.drive_gpu_stacked_pack();
-        let captured_at = self.session_start.map(|s| s.elapsed()).unwrap_or_default();
-        let rgba =
-            self.readback
-                .readback(self.pipeline.gpu(), self.pipeline.render_target(), cmd)?;
-        self.frame_count += 1;
-        if let (Some(replay), Some(bytes)) = (self.replay.as_mut(), rgba) {
-            replay.push(ReplayFrame {
-                rgba: bytes.to_vec(),
-                captured_at,
-                pose: ViewportPosition {
-                    yaw,
-                    pitch,
-                    fov_degrees: None,
-                },
-            });
-        }
-        Ok(match rgba {
-            Some(bytes) => RenderOutcome::Rgba(bytes),
-            None => RenderOutcome::Warmup,
-        })
-    }
-
-    /// Submit a stereo BGRA frame pair at an explicit pose. See
-    /// [`Self::submit_frame_yuv_at_pose`] for semantics.
-    ///
-    /// Does not run detection (BGRA backends are not yet supported;
-    /// see [`Self::submit_frame_bgra`] for the rationale).
-    pub fn submit_frame_bgra_at_pose(
-        &mut self,
-        left: &BgraPlanes<'_>,
-        right: &BgraPlanes<'_>,
-        yaw: f32,
-        pitch: f32,
-    ) -> Result<RenderOutcome<'_>, StitchCoreError> {
-        self.anchor_session_start();
-        let cmd = self
-            .pipeline
-            .render_to_target_bgra(left, right, yaw, pitch)?;
-        let captured_at = self.session_start.map(|s| s.elapsed()).unwrap_or_default();
-        let rgba =
-            self.readback
-                .readback(self.pipeline.gpu(), self.pipeline.render_target(), cmd)?;
-        self.frame_count += 1;
-        if let (Some(replay), Some(bytes)) = (self.replay.as_mut(), rgba) {
-            replay.push(ReplayFrame {
-                rgba: bytes.to_vec(),
-                captured_at,
-                pose: ViewportPosition {
-                    yaw,
-                    pitch,
-                    fov_degrees: None,
-                },
-            });
-        }
-        Ok(match rgba {
-            Some(bytes) => RenderOutcome::Rgba(bytes),
-            None => RenderOutcome::Warmup,
-        })
-    }
-
-    /// Submit a stereo packed-RGBA/BGRA frame pair and render the
-    /// current pose.
-    ///
-    /// Requires the core to have been built with [`InputFormat::Bgra`].
-    /// See [`Self::submit_frame_yuv`] for return semantics.
-    pub fn submit_frame_bgra(
-        &mut self,
-        left: &BgraPlanes<'_>,
-        right: &BgraPlanes<'_>,
-    ) -> Result<RenderOutcome<'_>, StitchCoreError> {
-        self.anchor_session_start();
-
-        // BGRA detection path: YOLO backends today consume YUV or
-        // NV12 `RawFrame` variants. Wrapping BGRA bytes as a YUV
-        // frame would require a color-space conversion we're not
-        // paying for yet - consumers that want detection on BGRA
-        // sources (OBS Browser Source, screen capture) attach a
-        // detector that understands BGRA once such a backend exists.
-        // For now, BGRA submits tick the director with the last
-        // detections (potentially from earlier YUV submits) but do
-        // not run detection themselves.
-
-        // `fresh_detection = false`: BGRA submits never run detection by
-        // design (see comment above). Directors must see this frame as
-        // "reusing cached detections" even on interval ticks, otherwise
-        // hysteresis counters over-fire.
-        let pose = self.resolve_current_pose(false);
-        let cmd = self
-            .pipeline
-            .render_to_target_bgra(left, right, pose.yaw, pose.pitch)?;
-        let captured_at = self.session_start.map(|s| s.elapsed()).unwrap_or_default();
-        let rgba =
-            self.readback
-                .readback(self.pipeline.gpu(), self.pipeline.render_target(), cmd)?;
-        self.frame_count += 1;
-        if let (Some(replay), Some(bytes)) = (self.replay.as_mut(), rgba) {
-            replay.push(ReplayFrame {
-                rgba: bytes.to_vec(),
-                captured_at,
-                pose,
-            });
-        }
-        Ok(match rgba {
-            Some(bytes) => RenderOutcome::Rgba(bytes),
-            None => RenderOutcome::Warmup,
-        })
-    }
-
-    /// Drain one pending readback slot without submitting a new frame.
-    ///
-    /// Useful at shutdown to collect the 1-2 frames still in-flight in
-    /// the triple-buffered staging pipeline.
-    pub fn flush(&mut self) -> Result<Option<&[u8]>, StitchCoreError> {
-        Ok(self.readback.flush_pending(self.pipeline.gpu())?)
-    }
-
-    // -----------------------------------------------------------------
-    // Low-level render-at-pose methods
-    //
-    // These produce a `wgpu::CommandBuffer` at a **caller-supplied pose**
-    // without running detection, without ticking the director, and
-    // without performing RGBA readback. They exist so consumers that
-    // need the rendered GPU texture as input to further GPU work
-    // (NV12 conversion for encoding, compositor texture import) can
-    // drive the core without paying for readback.
-    //
-    // The M3 `StitchSession::run` pull-adapter (plan step 2) uses these
-    // to route its encode loop through `StitchCore`: session owns its
-    // own director + detection pipeline during the transition and
-    // passes the resolved pose explicitly here. Once the session
-    // migration completes, these remain as the "render primitives" for
-    // multi-output consumers (record + stream, zero-copy compositor).
-    // -----------------------------------------------------------------
-
-    /// Render a stereo YUV420P frame at an explicit pose.
-    ///
-    /// Does not run detection, does not tick the director, does not
-    /// read back RGBA. Consumers that want the full `submit_*` loop
-    /// (detection + director + readback) should call
-    /// [`Self::submit_frame_yuv`] instead.
-    ///
-    /// The caller is responsible for subsequently consuming the
-    /// rendered texture (via [`Self::pipeline`] + `render_target()`)
-    /// or submitting the returned command buffer to chain further
-    /// GPU work.
-    pub fn render_yuv_at_pose(
-        &self,
-        left: &YuvPlanes<'_>,
-        right: &YuvPlanes<'_>,
-        yaw: f32,
-        pitch: f32,
-    ) -> Result<wgpu::CommandBuffer, StitchCoreError> {
-        Ok(self.pipeline.render_to_target(left, right, yaw, pitch)?)
-    }
-
-    /// Render a stereo packed-RGBA/BGRA frame at an explicit pose.
-    /// See [`Self::render_yuv_at_pose`] for semantics.
-    pub fn render_bgra_at_pose(
-        &self,
-        left: &BgraPlanes<'_>,
-        right: &BgraPlanes<'_>,
-        yaw: f32,
-        pitch: f32,
-    ) -> Result<wgpu::CommandBuffer, StitchCoreError> {
-        Ok(self
-            .pipeline
-            .render_to_target_bgra(left, right, yaw, pitch)?)
-    }
-
-    /// Render from GPU-resident RGBA textures (e.g. Bayer demosaic output).
-    ///
-    /// Copies the demosaiced textures into the stitch pipeline's input
-    /// planes (GPU-to-GPU blit), then renders the stitch. Returns the
-    /// render command buffer for `submit_render_output`.
-    pub fn render_gpu_rgba_at_pose(
-        &self,
-        left_rgba: &wgpu::Texture,
-        right_rgba: &wgpu::Texture,
-        yaw: f32,
-        pitch: f32,
-    ) -> wgpu::CommandBuffer {
-        self.pipeline
-            .render_from_gpu_rgba(left_rgba, right_rgba, yaw, pitch)
-    }
-
-    /// Render any [`StereoFrame`](crate::source::StereoFrame) variant
-    /// (YUV / NV12 / GpuResident) at an explicit pose.
-    ///
-    /// Thin wrapper over
-    /// [`StitchPipeline::render_stereo_frame`](crate::pipeline::StitchPipeline::render_stereo_frame)
-    /// that converts the pipeline error into a `StitchCoreError`. The
-    /// `MetalResident` variant is NOT handled here; use
-    /// [`Self::render_imported_textures_at_pose`] after importing the
-    /// `CVPixelBuffer` via `MetalTextureCache`.
-    pub fn render_stereo_frame_at_pose(
-        &self,
-        frame: &crate::source::StereoFrame,
-        yaw: f32,
-        pitch: f32,
-    ) -> Result<wgpu::CommandBuffer, StitchCoreError> {
-        Ok(self.pipeline.render_stereo_frame(frame, yaw, pitch)?)
-    }
-
-    /// Render from four pre-imported textures at an explicit pose.
-    ///
-    /// Used by the macOS zero-copy path where `CVPixelBuffer` Y/UV
-    /// planes are imported as wgpu textures via `MetalTextureCache`
-    /// (in `metal_interop`), and the Linux zero-copy path that shares
-    /// textures through the bind-group variant below.
-    pub fn render_imported_textures_at_pose(
-        &mut self,
-        left_y: &wgpu::Texture,
-        left_uv: &wgpu::Texture,
-        right_y: &wgpu::Texture,
-        right_uv: &wgpu::Texture,
-        yaw: f32,
-        pitch: f32,
-    ) -> wgpu::CommandBuffer {
-        self.pipeline
-            .render_imported_textures(left_y, left_uv, right_y, right_uv, yaw, pitch)
-    }
-
-    /// Render from pre-built GPU texture views at an explicit pose.
-    ///
-    /// Used by the D3D11VA zero-copy path where NV12 plane views are
-    /// created with `TextureAspect::Plane0` / `Plane1`.
-    pub fn render_imported_views_at_pose(
-        &mut self,
-        left_y: &wgpu::TextureView,
-        left_uv: &wgpu::TextureView,
-        right_y: &wgpu::TextureView,
-        right_uv: &wgpu::TextureView,
-        yaw: f32,
-        pitch: f32,
-    ) -> wgpu::CommandBuffer {
-        self.pipeline
-            .render_imported_views(left_y, left_uv, right_y, right_uv, yaw, pitch)
-    }
-
-    /// Render from pre-configured GPU bind groups and decode slots at
-    /// an explicit pose (Linux zero-copy path).
-    ///
-    /// Thin wrapper over
-    /// [`StitchPipeline::render_gpu_frame`](crate::pipeline::StitchPipeline::render_gpu_frame).
-    /// Consumers must have already called
-    /// [`StitchPipeline::configure_gpu_source`] via [`Self::pipeline_mut`].
-    #[cfg(target_os = "linux")]
-    pub fn render_gpu_frame_at_pose(
-        &mut self,
-        bind_groups: &crate::pipeline::GpuSourceBindGroups,
-        left_slot: u8,
-        right_slot: u8,
-        yaw: f32,
-        pitch: f32,
-    ) -> wgpu::CommandBuffer {
-        self.pipeline
-            .render_gpu_frame(bind_groups, left_slot, right_slot, yaw, pitch)
     }
 
     // -----------------------------------------------------------------
@@ -700,282 +302,6 @@ impl StitchCore {
     pub fn clear_detector(&mut self) {
         self.detector = None;
         self.last_detections.clear();
-    }
-
-    /// Attach a stacked-video replay recorder (M6.5 item 3, push
-    /// side).
-    ///
-    /// Every subsequent YUV submit feeds the recorder before
-    /// rendering. Errors inside the recorder are swallowed so a
-    /// failing recorder cannot break the live stitch output; the
-    /// recorder's own implementation is expected to log any
-    /// failure. See [`StackedReplayRecorder`] for the full
-    /// contract.
-    ///
-    /// Dropping an existing recorder via [`Self::clear_stacked_recorder`]
-    /// is required before attaching a new one; otherwise the old
-    /// recording is quietly abandoned.
-    pub fn set_stacked_recorder(&mut self, recorder: Box<dyn StackedReplayRecorder>) {
-        if self.stacked_recorder.is_some() {
-            log::warn!(
-                "StitchCore::set_stacked_recorder replacing an existing recorder; \
-                 call clear_stacked_recorder first to finalize the previous recording"
-            );
-        }
-        log::info!("StitchCore: stacked-video replay recorder attached");
-        self.stacked_recorder = Some(recorder);
-    }
-
-    /// Drop the current replay recorder, calling `finish` first so
-    /// the recording file is finalized. No-op if no recorder is
-    /// attached.
-    pub fn clear_stacked_recorder(&mut self) {
-        if let Some(mut recorder) = self.stacked_recorder.take() {
-            recorder.finish();
-            log::info!("StitchCore: stacked-video replay recorder detached");
-        }
-    }
-
-    /// Flush the replay recorder's buffered bytes to disk. Call
-    /// periodically (e.g. once per second from a timer) so a
-    /// concurrent reader sees recent frames. No-op if no recorder
-    /// is attached.
-    pub fn flush_stacked_recorder(&mut self) {
-        if let Some(ref mut recorder) = self.stacked_recorder {
-            recorder.flush();
-        }
-    }
-
-    /// Enable the GPU-pack replay path (M7 pivot item 1).
-    ///
-    /// Builds a [`YuvStackPacker`] sized for `layout` x `output_size`
-    /// and wires it into subsequent YUV submit calls. The packer's
-    /// source-format variant is derived from the pipeline's input
-    /// format so consumers don't risk a YUV/NV12 mismatch.
-    ///
-    /// Call [`Self::set_stacked_gpu_recorder`] to attach an
-    /// atlas-consuming sink (typically a
-    /// [`reco_io`](../../../reco_io/index.html) encoder) before the
-    /// first submit, or later - the pack still runs either way and
-    /// the first two submits are warmup.
-    ///
-    /// Emits one `log::info!` line stating the pack path has been
-    /// chosen (GPU), the tile dims, `N`, and the source format - so
-    /// the CPU vs GPU decision is never silent.
-    ///
-    /// Returns `StitchCoreError::Config` when the pipeline's input
-    /// format is BGRA (the pack shader only handles YUV420P / NV12)
-    /// or when the layout / output dims violate YUV420P alignment.
-    pub fn enable_gpu_stacked_replay(
-        &mut self,
-        layout: StackGridLayout,
-        output_size: OutputTileSize,
-    ) -> Result<(), StitchCoreError> {
-        let source_format = match self.pipeline.input_format() {
-            InputFormat::Yuv420p => SourceFormat::Yuv420p,
-            InputFormat::Nv12 => SourceFormat::Nv12,
-            InputFormat::Bgra => {
-                return Err(StitchCoreError::Config(
-                    "GPU stacked replay requires YUV420P or NV12 input; BGRA pipelines must use \
-                     the CPU replay-recording path"
-                        .into(),
-                ));
-            }
-        };
-        let packer = YuvStackPacker::new(self.pipeline.gpu(), layout, output_size, source_format)?;
-        let (atlas_w, atlas_h) = packer.atlas_dims();
-        log::info!(
-            "reco-core: replay pack path = GPU shader (tiles {}x{} out, N={}, atlas {}x{}, source_format={:?})",
-            output_size.width,
-            output_size.height,
-            layout.capacity(),
-            atlas_w,
-            atlas_h,
-            source_format,
-        );
-        if self.stacked_recorder.is_some() {
-            log::warn!(
-                "StitchCore::enable_gpu_stacked_replay: a CPU StackedReplayRecorder is also \
-                 attached; both paths will run and duplicate work. Clear one to avoid \
-                 redundant recording."
-            );
-        }
-        self.stacked_packer = Some(packer);
-        Ok(())
-    }
-
-    /// Disable the GPU-pack replay path and drop the packer.
-    /// Also calls `finish` on any attached GPU recorder so its file
-    /// is finalized. No-op when the path was not enabled.
-    pub fn disable_gpu_stacked_replay(&mut self) {
-        if self.stacked_packer.take().is_some() {
-            log::info!("StitchCore: GPU stacked replay disabled");
-        }
-        self.clear_stacked_gpu_recorder();
-    }
-
-    /// Attach a GPU-pack atlas recorder. Must be called after
-    /// [`Self::enable_gpu_stacked_replay`] for the pack output to
-    /// reach disk - without a recorder the packer still runs but
-    /// the readback bytes are dropped.
-    pub fn set_stacked_gpu_recorder(&mut self, recorder: Box<dyn StackedReplayGpuRecorder>) {
-        if self.stacked_gpu_recorder.is_some() {
-            log::warn!(
-                "StitchCore::set_stacked_gpu_recorder replacing an existing GPU recorder; \
-                 call clear_stacked_gpu_recorder first to finalize the previous recording"
-            );
-        }
-        if self.stacked_packer.is_none() {
-            log::warn!(
-                "StitchCore::set_stacked_gpu_recorder called before \
-                 enable_gpu_stacked_replay: recorder will receive no atlases until the \
-                 packer is enabled"
-            );
-        }
-        log::info!("StitchCore: GPU stacked-replay recorder attached");
-        self.stacked_gpu_recorder = Some(recorder);
-    }
-
-    /// Drop the GPU-pack atlas recorder, calling `finish` so the
-    /// output file is finalized. No-op if no recorder is attached.
-    pub fn clear_stacked_gpu_recorder(&mut self) {
-        if let Some(mut recorder) = self.stacked_gpu_recorder.take() {
-            recorder.finish();
-            log::info!("StitchCore: GPU stacked-replay recorder detached");
-        }
-    }
-
-    /// Flush the GPU recorder's buffered bytes to disk. No-op if no
-    /// recorder is attached.
-    pub fn flush_stacked_gpu_recorder(&mut self) {
-        if let Some(ref mut recorder) = self.stacked_gpu_recorder {
-            recorder.flush();
-        }
-    }
-
-    /// Atlas dimensions `(width, height)` the current packer produces,
-    /// or `None` when GPU stacked replay is not enabled. Consumers
-    /// use this to open an encoder sized for the atlas.
-    pub fn stacked_atlas_dims(&self) -> Option<(u32, u32)> {
-        self.stacked_packer.as_ref().map(|p| p.atlas_dims())
-    }
-
-    /// Pack the GPU stacked-replay atlas from external texture
-    /// views (the zero-copy entry point).
-    ///
-    /// Used by session-layer zero-copy submit paths where source
-    /// frames live in shared / imported textures rather than the
-    /// renderer's internal plane textures. Call after the stitch
-    /// submit has landed; this method encodes a separate command
-    /// buffer for the pack + staging copy, submits it, and polls
-    /// the triple-buffer ring for a ready atlas to feed to the
-    /// attached recorder.
-    ///
-    /// The storytelling flow (per the project principle - no silent
-    /// decisions): the caller chose this path because the source is
-    /// GPU-resident. The packer's configured `SourceFormat` was
-    /// logged at `enable_gpu_stacked_replay` time. From here on,
-    /// every call is just bytes moving through the pipeline, so no
-    /// per-frame logging.
-    ///
-    /// No-op when the packer isn't enabled.
-    ///
-    /// Hard-coded to the two-camera stereo layout today; extend
-    /// when `CameraInput::camera_count() > 2` lands.
-    pub fn pack_gpu_stacked_replay_from_views(
-        &mut self,
-        left: StackedPackSource<'_>,
-        right: StackedPackSource<'_>,
-    ) {
-        crate::profile_scope!("replay_pack_from_views");
-        let Some(ref mut packer) = self.stacked_packer else {
-            return;
-        };
-        let gpu = self.pipeline.gpu();
-        let mut encoder = gpu
-            .device
-            .create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                label: Some("stitch_core_gpu_stacked_pack_ext"),
-            });
-        let capacity = packer.layout().capacity();
-        if capacity >= 1 {
-            packer.pack_tile_from_views(gpu, &mut encoder, 0, left);
-        }
-        if capacity >= 2 {
-            packer.pack_tile_from_views(gpu, &mut encoder, 1, right);
-        }
-        {
-            crate::profile_scope!("replay_copy_to_staging");
-            packer.copy_to_staging(&mut encoder);
-            gpu.queue.submit(Some(encoder.finish()));
-        }
-
-        {
-            crate::profile_scope!("replay_poll_and_record");
-            if let Some(atlas) = packer.poll_ready(gpu)
-                && let Some(ref mut recorder) = self.stacked_gpu_recorder
-            {
-                recorder.record_atlas(&atlas);
-            }
-        }
-    }
-
-    /// Runs the GPU pack from the pipeline's internal plane
-    /// textures. Used by CPU-upload submit paths where
-    /// `queue.write_texture` has just populated the renderer's
-    /// own textures - fires from `submit_frame_yuv*` and from the
-    /// session's `process_frame` non-zero-copy branch. Zero-copy
-    /// paths take [`Self::pack_gpu_stacked_replay_from_views`]
-    /// instead because their source data lives in shared
-    /// textures that bypass the renderer's internal planes.
-    ///
-    /// Delegates through the same pack + poll + record path so
-    /// every entry point shares behavior.
-    ///
-    /// No-op when the packer isn't enabled.
-    pub(crate) fn drive_gpu_stacked_pack(&mut self) {
-        crate::profile_scope!("replay_drive_pack");
-        if self.stacked_packer.is_none() {
-            return;
-        }
-        // Pipeline's plane-view accessors return
-        // (y_view, u_or_uv_view, v_or_dummy_view). Build the
-        // StackedPackSource variant matching the packer's
-        // configured source format - the packer will route to the
-        // right shader kernel internally.
-        let (ly, lu, lv) = self.pipeline.left_plane_views();
-        let (ry, ru, rv) = self.pipeline.right_plane_views();
-        // Keep bindings alive across the pack call via locals.
-        let (left, right) = match self.pipeline.input_format() {
-            InputFormat::Yuv420p => (
-                StackedPackSource::Yuv420p {
-                    y: &ly,
-                    u: &lu,
-                    v: &lv,
-                },
-                StackedPackSource::Yuv420p {
-                    y: &ry,
-                    u: &ru,
-                    v: &rv,
-                },
-            ),
-            InputFormat::Nv12 => (
-                StackedPackSource::Nv12 { y: &ly, uv: &lu },
-                StackedPackSource::Nv12 { y: &ry, uv: &ru },
-            ),
-            InputFormat::Bgra => {
-                // Shouldn't happen: enable_gpu_stacked_replay
-                // rejects BGRA up front. Defensive no-op so the
-                // live render loop can't panic on an invariant
-                // violation.
-                log::error!(
-                    "drive_gpu_stacked_pack: packer enabled but pipeline input_format is \
-                     BGRA; skipping pack (this is a logic bug in StitchCore)"
-                );
-                return;
-            }
-        };
-        self.pack_gpu_stacked_replay_from_views(left, right);
     }
 
     /// Set how often detection runs.
@@ -1117,11 +443,6 @@ impl StitchCore {
     }
 
     // -----------------------------------------------------------------
-    // Pipeline stages (slot reserved; no registered stages yet)
-    // -----------------------------------------------------------------
-
-
-    // -----------------------------------------------------------------
     // Replay buffer
     // -----------------------------------------------------------------
 
@@ -1178,139 +499,15 @@ impl StitchCore {
     pub fn gpu(&self) -> &GpuContext {
         self.pipeline.gpu()
     }
-
-    // -----------------------------------------------------------------
-    // Internals
-    // -----------------------------------------------------------------
-
-    fn anchor_session_start(&mut self) {
-        if self.session_start.is_none() {
-            self.session_start = Some(Instant::now());
-        }
-    }
-
-    fn resolve_current_pose(&mut self, fresh_detection: bool) -> ViewportPosition {
-        // Pull raw director output (or default) and clamp through
-        // coverage. Then write the resolved FOV back onto the pipeline
-        // so the upcoming render uses it.
-        //
-        // `fresh_detection` is the ACTUAL run decision for this frame,
-        // not the schedule-would-fire predicate. The BGRA submit path
-        // deliberately skips detection (no BGRA-aware backend exists
-        // today) so it must pass `false` even when the interval would
-        // have fired - otherwise directors over-count hysteresis on
-        // stale cached detections.
-        let timestamp_ms = self
-            .session_start
-            .map(|s| s.elapsed().as_secs_f64() * 1000.0)
-            .unwrap_or(0.0);
-
-        // Pose resolution: run all registered trackers in order, assemble
-        // a [`WorldState`], and let the panner decide via the shared
-        // `panner::dispatch` helper. When no panner is attached the pose
-        // stays at the pipeline default.
-        let _ = fresh_detection; // reserved for future freshness-aware panners
-        let raw = crate::panner::dispatch(
-            self.panner.as_mut(),
-            self.player_tracker.as_mut(),
-            self.ball_tracker.as_mut(),
-            &mut self.previous_panner_pose,
-            // StitchCore does not own an event sink. StitchSession
-            // does the tracing when it is the active entry point.
-            None,
-            crate::panner::DispatchContext {
-                detections: &self.last_detections,
-                calibration: &self.pipeline.calibration,
-                frame_index: self.frame_count,
-                timestamp_ms,
-                caller: "StitchCore",
-            },
-        )
-        .map(|r| r.pose)
-        .unwrap_or_default();
-        let clamped = if self.constrained_look {
-            self.safe_clamp(raw)
-        } else {
-            raw
-        };
-        if let Some(fov) = clamped.fov_degrees {
-            self.pipeline.set_fov(fov);
-        }
-        clamped
-    }
-
-    fn should_run_detection(&self) -> bool {
-        // Before any submit, frame_count == 0 and interval defaults to
-        // 1, which covers the "run on the very first frame" case.
-        self.detection_interval > 0 && self.frame_count.is_multiple_of(self.detection_interval)
-    }
-
-    /// Run the attached detector against a stereo YUV420P frame pair.
-    ///
-    /// Wraps each plane as [`RawFrame`] + [`DetectorFrame::Cpu`] and
-    /// dispatches through the unified trait, once per camera. Errors
-    /// are warned-and-dropped: a flaky inference call must not crash
-    /// the render loop.
-    fn run_yuv_detection(
-        &mut self,
-        left: &YuvPlanes<'_>,
-        right: &YuvPlanes<'_>,
-        source_width: u32,
-        source_height: u32,
-    ) -> Vec<Detection> {
-        let Some(ref mut detector) = self.detector else {
-            return Vec::new();
-        };
-        let mut out = Vec::new();
-        for (camera, planes) in [(CameraId::Left, left), (CameraId::Right, right)] {
-            let raw = RawFrame {
-                y: planes.y,
-                chroma: ChromaFormat::Yuv420p {
-                    u: planes.u,
-                    v: planes.v,
-                },
-                width: source_width,
-                height: source_height,
-            };
-            match detector.detect(camera, &DetectorFrame::Cpu(raw)) {
-                Ok(v) => out.extend(v),
-                Err(e) => log::warn!("StitchCore detector '{}' {camera:?}: {e}", detector.name()),
-            }
-        }
-        out
-    }
-
-    /// Map raw camera-space detections to panorama-space
-    /// [`MappedDetection`]s the director can consume.
-    fn map_detections_to_panorama(&self, detections: Vec<Detection>) -> Vec<MappedDetection> {
-        let calibration = self.pipeline.calibration();
-        let scene = &self.pipeline.scene;
-        detections
-            .into_iter()
-            .map(|d| {
-                let position = projection::camera_to_panorama(
-                    d.camera,
-                    d.center_x,
-                    d.center_y,
-                    calibration,
-                    scene,
-                );
-                MappedDetection {
-                    camera: d.camera,
-                    class_id: d.class_id,
-                    confidence: d.confidence,
-                    camera_center: (d.center_x, d.center_y),
-                    camera_size: (d.width, d.height),
-                    position,
-                }
-            })
-            .collect()
-    }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use std::time::Duration;
+
+    use crate::core::replay_buffer::ReplayBuffer;
+    use crate::core::types::{RenderOutcome, ReplayFrame, StitchCoreError};
+    use crate::director::ViewportPosition;
 
     /// Assert `ReplayBuffer` trims old frames as the newest ages past
     /// `max_duration`. This is the core guarantee OBS A16 relies on:
